try to solve some problems
I found many problems when I was learning Hedy at the Hedy website (Chinese version), so I made some changes to correct the programming and I will check it again.  I found it is difficult to use a space correctly when I use both Chinese characters and English characters in a sentence because we don't use a space between each Chinese character or between a Chinese character and Chinese punctuation. I think we should mention these rules of using punctuation for Chinese students.